requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tweepy>=4.14.0
PyGithub>=2.1.1
PyPDF2>=3.0.0
openai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0

